+++ src/models/StyledComponent.js
@@ -2,6 +2,10 @@
 
 import { createElement } from 'react'
 
+import type { Theme } from './ThemeProvider'
+
+import type { Theme } from './ThemeProvider'
+
 import validAttr from '../utils/validAttr'
 import isTag from '../utils/isTag'
 import type { RuleSet, Target } from '../types'
@@ -19,7 +23,10 @@ const safeDisplayName = (displayName: string) => (
 
 export default (ComponentStyle: Function) => {
   // eslint-disable-next-line no-undef
-  const createStyledComponent = (target: Target, rules: RuleSet, parent?: ReactClass<*>) => {
+  const createStyledComponent = (input: Target, rules: RuleSet, parent?: ReactClass<*>) => {
+    const target = input.target || input
+    const { identifier, passedDisplayName } = input
+
     /* Handle styled(OtherStyledComponent) differently */
     const isStyledComponent = AbstractStyledComponent.isPrototypeOf(target)
     if (!isTag(target) && isStyledComponent) {
@@ -37,7 +44,7 @@ export default (ComponentStyle: Function) => {
       constructor() {
         super()
         this.state = {
-          theme: {},
+          theme: null,
           generatedClassName: '',
         }
       }
@@ -59,20 +66,22 @@ export default (ComponentStyle: Function) => {
             this.setState({ theme, generatedClassName })
           })
         } else {
+          const theme = this.props.theme || {}
           const generatedClassName = this.generateAndInjectStyles(
-            this.props.theme || {},
+            theme,
             this.props
           )
-          this.setState({ generatedClassName })
+          this.setState({ theme, generatedClassName })
         }
       }
 
-      componentWillReceiveProps(nextProps: any) {
-        const generatedClassName = this.generateAndInjectStyles(
-          this.state.theme || this.props.theme,
-          nextProps
-        )
-        this.setState({ generatedClassName })
+      componentWillReceiveProps(nextProps: { theme?: Theme, [key: string]: any }) {
+        this.setState((oldState) => {
+          const theme = nextProps.theme || oldState.theme
+          const generatedClassName = this.generateAndInjectStyles(theme, nextProps)
+
+          return { theme, generatedClassName }
+        })
       }
 
       componentWillUnmount() {
@@ -101,8 +110,12 @@ export default (ComponentStyle: Function) => {
       }
     }
 
-    let displayName = isTag ? `styled.${target}` : `Styled(${target.displayName})`
+    let displayName = passedDisplayName || isTag ? `styled.${target}` : `Styled(${target.displayName})`
     const generateIdentifierFromDisplayName = (includeDisplayName: boolean) => {
+      if (identifier) {
+        StyledComponent.identifier = identifier
+        return
+      }
       const nr = (identifiers[displayName] || 0) + 1
       identifiers[displayName] = nr
       const hash = componentStyle.generateName(displayName + nr)
@@ -115,7 +128,6 @@ export default (ComponentStyle: Function) => {
         return displayName
       },
       set(newName) {
-        console.log('SETTING DISPLAY NAME')
         displayName = newName
         generateIdentifierFromDisplayName(true)
       },