+++ src/test/basic.test.js
@@ -1,3 +1,4 @@
+// @flow
 import React, { Component } from 'react'
 import expect from 'expect'
 import { shallow, mount } from 'enzyme'
@@ -34,21 +35,36 @@ describe('basic', () => {
   it('should generate an empty tag once rendered', () => {
     const Comp = styled.div``
     shallow(<Comp />)
-    expectCSSMatches('.a {  }')
+    expectCSSMatches('.b {  }')
   })
 
   /* TODO: we should probably pretty-format the output so this test might have to change */
   it('should pass through all whitespace', () => {
     const Comp = styled.div`   \n   `
     shallow(<Comp />)
-    expectCSSMatches('.a {    \n    }', { skipWhitespace: false })
+    expectCSSMatches('.b {    \n    }', { ignoreWhitespace: false })
   })
 
   it('should inject only once for a styled component, no matter how often it\'s mounted', () => {
     const Comp = styled.div``
     shallow(<Comp />)
     shallow(<Comp />)
-    expectCSSMatches('.a {  }')
+    expectCSSMatches('.b {  }')
+  })
+
+  describe('private API', () => {
+    it('should accept the private, internal API notation', () => {
+      const Comp = styled({ target: 'div', identifier: 'Comp-a123bf', displayName: 'Comp' })``
+      shallow(<Comp />)
+      expectCSSMatches('.a { }')
+    })
+
+    it('should add the passed identifier as a class', () => {
+      const identifier = 'Comp-a123bf'
+      const Comp = styled({ target: 'div', identifier: identifier, displayName: 'Comp' })``
+      const renderedComp = shallow(<Comp />)
+      expect(renderedComp.hasClass(identifier)).toBe(true)
+    })
   })
 
   describe('jsdom tests', () => {
@@ -56,12 +72,14 @@ describe('basic', () => {
     it('should pass ref to the component', () => {
       const Comp = styled.div``
       const WrapperComp = class extends Component {
+        testRef: any;
         render() {
           return <Comp innerRef={(comp) => { this.testRef = comp }} />
         }
       }
 
       const wrapper = mount(<WrapperComp />)
+      // $FlowIssue
       expect(wrapper.node.testRef).toExist()
     })
   })