+++ src/test/extending.test.js
@@ -1,3 +1,4 @@
+// @flow
 import React from 'react'
 import expect from 'expect'
 import { shallow } from 'enzyme'
@@ -14,34 +15,34 @@ describe('extending', () => {
     styled = resetStyled()
   })
 
-  it('should generate a single class with no styles', () => {
+  it('should generate two classes with no styles', () => {
     const Parent = styled.div``
     const Child = styled(Parent)``
 
     shallow(<Parent />)
     shallow(<Child />)
 
-    expectCSSMatches('.a { }')
+    expectCSSMatches('.c { } .d { }')
   })
 
-  it('should generate a single class if only parent has styles', () => {
+  it('should generate two classes if only parent has styles', () => {
     const Parent = styled.div`color: blue;`
     const Child = styled(Parent)``
 
     shallow(<Parent />)
     shallow(<Child />)
 
-    expectCSSMatches('.a { color: blue; }')
+    expectCSSMatches('.b { color: blue; } .c { }')
   })
 
-  it('should generate a single class if only child has styles', () => {
+  it('should generate two classes if only child has styles', () => {
     const Parent = styled.div`color: blue;`
     const Child = styled(Parent)``
 
     shallow(<Parent />)
     shallow(<Child />)
 
-    expectCSSMatches('.a { color: blue; }')
+    expectCSSMatches('.b { color: blue; } .c { }')
   })
 
   it('should generate a class for the child with the rules of the parent', () => {