+++ src/types.js
@@ -1,9 +1,12 @@
 // @flow
-export type Interpolation = ((executionContext: Object) => string) | string
+export type Interpolation = ((executionContext: Object) => string) | string | number
 /* todo: I want this to actually be an array of Function | string but that causes errors */
 export type RuleSet = Array<Interpolation>
 
 /* eslint-disable no-undef */
-export type Target = string | ReactClass<*>
+type Tag = string | ReactClass<*>
+type Input = { tag: Tag, identifier?: string, displayName?: string }
+
+export type Target = Tag | Input
 
 export type NameGenerator = (hash: number) => string